requirements.txt
dbt-core
dbt-duckdb
flask